Made following changes in volume metrics code
Introduced slabs for all hash table entries.
Periodically the volume name cache will be dropped.
Timestamp records must be merged from different files before pushing to tsdb.
Change the algorithm in the way we merge timestamp records and push to tsdb.
Changed the algorithm in the way we remember the current offset and dispatch
to mfs (xattr).