MAPRDB-1226: tweak behavior of max bucket of table metrics latencies
Currently catch-all bucket is  10s--infinity. From now on *for
purposes of publishing* we show the catch-all buchet as 10s--10.001s
so that percentile calculation would bever go above 10.001s. Monet
will document that if latencies are above 10s, the only thing we
know about them is that they are above 10s (and we don't know how
much exactly are they).